+++ support/updatePodfile.ts
@@ -8,7 +8,7 @@ export async function updatePodfile(iosPath: string) {
   const matches = podfile.match(NSE_PODFILE_REGEX);
 
   if (matches) {
-    OneSignalLog.log("OneSignalNotificationServiceExtension target already added to Podfile. Skipping...");
+    OneSignalLog.log("BituNotificationServiceExtension target already added to Podfile. Skipping...");
   } else {
     fs.appendFile(`${iosPath}/Podfile`, NSE_PODFILE_SNIPPET, (err) => {
       if (err) {